:recycle: Create header component to render correct one in each route
